add run script and instructions, fix closure and print statement
