+++ RUN_INSTRUCTIONS.md
@@ -1,0 +1,3 @@
+###Run Instructions###
+1) From a Unix command line, enter the root directory for this project.
+2) Run the RUN_TESTS.sh shell script via "./RUN_TESTS.sh".
