+++ RUN_TESTS.sh
@@ -1,0 +1,2 @@
+cd test_cases
+python ../test_runner.py "python3 ../src/SmurfInvoker.py"
